requirements.txt
replicate
requests
Pillow
python-dotenv
tkinter
